rag/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
sentence-transformers
chromadb
google-generativeai
numpy
